r-band at rawseeing=0.5 arcsec
Data are in 10210238/.  Data are the centroid file cointaining
the pixel position and flux of each source and the image produced for each chip.

The input file is 10210238_f2_0.5.pars.  The extra commands file is
nobackground.  An example of a string used to run phosim is:
>>> phosim 10210238_f2_0.5.pars -s 'R22_S11' -c examples/nobackground -o output

These were generated with brighter fatter turned off.