i-band at rawseeing=0.8 arcsec
Data are in 23983043/.  Data are the centroid file cointaining
the pixel position and flux of each source and the image produced for each chip.

I had problems with saturation with this particular visit.  I changed the magNorm
to 16.0 for this and all subsequent visits.

I also turned off the pixel boundary uncertainties (pixelerror 0).  This had been done
for the last visit as well.  See the nobackground file for the physics parameters invoked
for each visit.

The input file is i23983043_f3_0.8.pars.  The extra commands file is
nobackground.  An example of a string used to run phosim is:
>>> phosim 23983043_f3_0.8.pars -s 'R22_S11' -c examples/nobackground -o output