i-band at rawseeing=0.7 arcsec
Data are in 33235904/.  Data are the centroid file cointaining
the pixel position and flux of each source and the image produced for each chip.

This visit has clouds turned off.  It also has saturation and blooming turned off.
This means that from this visit on, no PSFs should be culled for having pixels over
100000.

The input file is 33235904_f3_0.7.pars.  The extra commands file is
nobackground.  An example of a string used to run phosim is:
>>> phosim 10210238_f2_0.5.pars -s 'R22_S11' -c examples/nobackground -o output